--- a/cabal.project
+++ b/cabal.project
@@ -1,3 +1,4 @@
 packages: webgear-core/
           webgear-server/
           webgear-openapi/
+          webgear-benchmarks/