--- a/stack-8.10.7.yaml
+++ b/stack-8.10.7.yaml
@@ -4,3 +4,4 @@ packages:
   - ./webgear-core
   - ./webgear-server
   - ./webgear-openapi
+  - ./webgear-benchmarks