--- a/webgear-benchmarks/webgear-benchmarks.cabal
+++ b/webgear-benchmarks/webgear-benchmarks.cabal
@@ -19,7 +19,7 @@ source-repository head
 
 executable users
   default-language:   Haskell2010
-  build-depends:      aeson ==2.0.3.0
+  build-depends:      aeson >=1.5 && <2.1
                     , base >=4.14.0.0 && <5
                     , bytestring >=0.10 && <0.12
                     , criterion ==1.5.*
@@ -28,7 +28,7 @@ executable users
                     , mtl ==2.2.2
                     , text ==1.2.*
                     , time
-                    , unordered-containers ==0.2.17.0
+                    , unordered-containers ==0.2.*
                     , servant ==0.19
                     , servant-server ==0.19.1
                     , scotty ==0.12