… .github/workflows/ci.yaml
… hie.yaml
… stack-9.0.2.yaml
… stack-9.2.3.yaml
… webgear-benchmarks/README.md
… webgear-benchmarks/src/users/WebGear.hs
… webgear-benchmarks/user.json
… webgear-core/src/WebGear/Core/Trait/Auth/Basic.hs
… webgear-core/src/WebGear/Core/Trait/Auth/Common.hs
… webgear-core/src/WebGear/Core/Trait/Auth/JWT.hs
… webgear-core/src/WebGear/Core/Trait/Body.hs
… webgear-core/src/WebGear/Core/Trait/Header.hs
… webgear-core/src/WebGear/Core/Trait/Path.hs
… webgear-core/src/WebGear/Core/Trait/QueryParam.hs
… webgear-core/src/WebGear/Core/Trait/Status.hs
… webgear-openapi/src/WebGear/OpenApi/Handler.hs
… webgear-openapi/src/WebGear/OpenApi/Trait/Auth/Basic.hs
… webgear-openapi/src/WebGear/OpenApi/Trait/Auth/JWT.hs
… webgear-openapi/src/WebGear/OpenApi/Trait/Body.hs
… webgear-openapi/src/WebGear/OpenApi/Trait/Header.hs
… webgear-openapi/src/WebGear/OpenApi/Trait/Method.hs
… webgear-openapi/src/WebGear/OpenApi/Trait/Path.hs
… webgear-openapi/src/WebGear/OpenApi/Trait/QueryParam.hs
… webgear-openapi/src/WebGear/OpenApi/Trait/Status.hs
… webgear-server/src/WebGear/Server/Handler.hs
… webgear-server/src/WebGear/Server/Trait/Auth/Basic.hs
… webgear-server/src/WebGear/Server/Trait/Auth/JWT.hs
… webgear-server/src/WebGear/Server/Trait/Body.hs
… webgear-server/src/WebGear/Server/Trait/Header.hs
… webgear-server/src/WebGear/Server/Trait/Method.hs