WebGear has a performance degradation in GHC 9.2.x compared to previous GHC versions. It looks like servant and scotty get optimized better in 9.2.x compared to WebGear.

Adding INLINE pragmas brings benchmark numbers closer to servant and scotty.

Sample benchmark:

```
Run GHC_VERSION=$(echo 9.2.4 | tr -d .)
  GHC_VERSION=$(echo 9.2.4 | tr -d .)
  BENCHMARK=$(nix build --no-link --json .#webgear-benchmarks-ghc${GHC_VERSION} | jq -r .[].outputs.out)
  ${BENCHMARK}/bin/users +RTS -N -RTS
  shell: /usr/bin/bash -e {0}
benchmarking webgear/500
time                 10.57 ms   (10.53 ms .. 10.63 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 10.58 ms   (10.54 ms .. 10.64 ms)
std dev              112.5 μs   (78.97 μs .. 158.0 μs)

benchmarking webgear/1000
time                 21.17 ms   (21.07 ms .. 21.32 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 21.09 ms   (21.04 ms .. 21.19 ms)
std dev              165.3 μs   (68.31 μs .. 268.5 μs)

benchmarking webgear/5000
time                 105.9 ms   (105.2 ms .. 106.6 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 105.9 ms   (105.7 ms .. 106.3 ms)
std dev              423.9 μs   (253.2 μs .. 615.2 μs)

benchmarking servant/500
time                 8.338 ms   (8.329 ms .. 8.352 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 8.335 ms   (8.329 ms .. 8.344 ms)
std dev              22.53 μs   (15.68 μs .. 36.11 μs)

benchmarking servant/1000
time                 16.76 ms   (16.68 ms .. 16.83 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 16.75 ms   (16.71 ms .. 16.79 ms)
std dev              107.7 μs   (60.38 μs .. 156.1 μs)

benchmarking servant/5000
time                 83.39 ms   (83.09 ms .. 83.59 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 83.42 ms   (83.33 ms .. 83.55 ms)
std dev              177.9 μs   (133.6 μs .. 238.5 μs)

benchmarking scotty/500
time                 10.19 ms   (10.17 ms .. 10.21 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 10.18 ms   (10.17 ms .. 10.22 ms)
std dev              64.47 μs   (18.69 μs .. 128.4 μs)

benchmarking scotty/1000
time                 20.19 ms   (20.05 ms .. 20.27 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 20.31 ms   (20.24 ms .. 20.55 ms)
std dev              278.3 μs   (51.45 μs .. 537.0 μs)

benchmarking scotty/5000
time                 101.7 ms   (101.0 ms .. 102.9 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 101.6 ms   (101.3 ms .. 102.0 ms)
std dev              546.6 μs   (252.2 μs .. 817.6 μs)
```